+++ CHANGELOG.md
@@ -5,12 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.8.0] - 2019-12-21
 ### Added
 - `Error::MalformedEntityReference`
 
 ### Changed
 - Malformed entity reference is an error now.
 - Escaped `<` in attribute inside an ENTITY is an error now.
+- `xmlparser` updated with multiple fixes.
 
 ## [0.7.3] - 2019-11-14
 ### Changed
@@ -102,7 +105,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `ExpandedName::has_namespace`. `ExpandedName::namespace` will return `Option` now.
 - `Error::NestedEntityReference`.
 
-[Unreleased]: https://github.com/RazrFalcon/roxmltree/compare/v0.7.3...HEAD
+[Unreleased]: https://github.com/RazrFalcon/roxmltree/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/RazrFalcon/roxmltree/compare/v0.7.3...v0.8.0
 [0.7.3]: https://github.com/RazrFalcon/roxmltree/compare/v0.7.2...v0.7.3
 [0.7.2]: https://github.com/RazrFalcon/roxmltree/compare/v0.7.1...v0.7.2
 [0.7.1]: https://github.com/RazrFalcon/roxmltree/compare/v0.7.0...v0.7.1